Our base has collisions now <3 Also added some more objects in Level_1 and upgraded Translator script (it's more universal now and remembers its direction and speed for specific objects)

+++ PBLgame/PBLgame/Engine/Physics/SphereCollider.cs
@@ -361,7 +361,7 @@ namespace PBLgame.Engine.Physics
         {
             CultureInfo culture = CultureInfo.InvariantCulture;
 
-            Radius = Convert.ToInt32(reader.GetAttribute("Radius"), culture);
+            Radius = Convert.ToSingle(reader.GetAttribute("Radius"), culture);
             Trigger = Convert.ToBoolean(reader.GetAttribute("IsTrigger"), culture);
             reader.ReadStartElement();
             if (reader.Name == "LocalPosition")
